+++ src/Orgabat/GameBundle/Entity/Category.php
@@ -109,17 +109,20 @@ class Category
         return $this->exercises;
     }
 
+    /**
+     * Get the sum of all best exercises for a category
+     *
+     * @return array : results of all exercises for a category
+     */
     public function getDataFromAllBestExercises() {
         $results = ["timer" => 0, "healthNote" => 0, "organizationNote" => 0, "businessNotorietyNote" => 0];
-        $compteur = 0;
         foreach ($this->exercises as $exercise) {
             $results["timer"] += $exercise->getBestExerciseHistory()->getTimer();
             $results["healthNote"] += $exercise->getBestExerciseHistory()->getHealthNote();
             $results["organizationNote"] += $exercise->getBestExerciseHistory()->getOrganizationNote();
             $results["businessNotorietyNote"] += $exercise->getBestExerciseHistory()->getBusinessNotorietyNote();
-            $compteur ++;
         }
-        $results["timer"] /= $compteur;
+        $results["timer"] /= count($this->exercises);
         return $results;
     }
 }
