+++ src/Orgabat/GameBundle/Resources/views/Pdf/allUsers.html.twig
@@ -54,6 +54,8 @@
 <div class="container">
     <h1 class="">Orgabat</h1>
     <h2 class="">Résultats des jeux</h2>
+
+    {# User informations #}
     {% for  key, user in users %}
         <div class="user-info">
             <p class="name">{{ user.lastName|upper }} {{ user.firstName }}</p>
@@ -62,9 +64,13 @@
                 <p class="section">Inscrit en <span class="sectionName">{{ user.section.name }}</span></p>
             {% endif %}
         </div>
+
+        {# If an apprentice didnt played #}
         {% if categories[key] is empty %}
             <h3>Pas de données de jeu enregistrées.</h3>
         {% else %}
+
+            {# Iterate all categories played by an apprentice #}
             {% for category in categories[key]  %}
                 <h3 class=" titlePdf">{{ category.name }}</h3>
                 <table class="table-stats">
@@ -87,6 +93,8 @@
                             <td>{{ exercise.bestExerciseHistory.businessNotorietyNote }} / {{ exercise.businessNotorietyMaxNote }}</td>
                         </tr>
                     {% endfor %}
+
+                    {# Total stats for a category #}
                     <tr>
                         <td>Total</td>
                         <td>{{ category.dataFromAllBestExercises.timer }}</td>
